make dev: boot embedded PGlite + seed dev login when DATABASE_URL is unset
- dev-native.sh: no DATABASE_URL → `bun run @lobu/server dev:local` (start-local.ts) instead of erroring; mints dev@lobu.local / lobudev123 + bootstrap PAT
- extract shared mountViteDev() into dev-vite.ts so start-local.ts also serves the SPA with HMR
- start-local.ts: also look for db/migrations at the monorepo root (tsx run, not the published bundle)
