+++ packages/server/package.json
@@ -10,6 +10,7 @@
   },
   "scripts": {
     "dev": "tsx watch --ignore=../web/** --ignore=../../node_modules/** src/server.ts",
+    "dev:local": "tsx watch --ignore=../web/** --ignore=../../node_modules/** src/start-local.ts",
     "start": "tsx src/server.ts",
     "build:server": "node ./scripts/build-server-bundle.mjs",
     "test": "vitest",
